Backend/apps/reminders/serializers.py
from rest_framework import serializers
from .models import Reminder

class ReminderSerializer(serializers.ModelSerializer):
    class Meta:
        model = Reminder
        fields = ['id', 'application', 'reminder_date', 'message']
        read_only_fields = ['user'] 